% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_cardinal_vars}
\alias{get_cardinal_vars}
\title{Returns cardinality information of the dataframe - The cardinalities for the variables (applies to factor variables)}
\usage{
get_cardinal_vars(df,
  cardinality_threshold = cardinality_threshold_pkg_default)
}
\arguments{
\item{df}{The dataframe to generate the EDA report}

\item{cardinality_threshold}{OPTIONAL. The threshold based on which variables will be flagged as "High Cardinality". Default value is 50}
}
\value{
cardinalities The output dataframe containing the column name and the cardinality
}
\description{
Returns cardinality information of the dataframe - The cardinalities for the variables (applies to factor variables)
}
\examples{
get_cardinal_vars(df)
get_cardinal_vars(df, cardinality_threshold=0.3)
}
