% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_correlated_vars}
\alias{get_correlated_vars}
\title{Returns skewness information of the dataframe}
\usage{
get_correlated_vars(df, correlation_threshold = correlation_threshold,
  method_name = "pearson")
}
\arguments{
\item{df}{The dataframe to generate the EDA report}

\item{correlation_threshold}{OPTIONAL. The threshold based on which variables will be flagged as "Highly Correlated". Default value is 0.8}
}
\value{
melted_cormat_f_one A dataframe containing the correlated variables and their corresponding pearson correlation coefficient
}
\description{
Returns skewness information of the dataframe
}
\examples{
get_correlated_vars(df)
get_correlated_vars(df, correlation_threshold=20)
}
