% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_full_warnings}
\alias{get_full_warnings}
\title{Returns skewness information of the dataframe}
\usage{
get_full_warnings(df, missing_threshold = missing_threshold_pkg_default,
  cardinality_threshold = cardinality_threshold_pkg_default,
  skewness_threshold = skewness_threshold_pkg_default,
  correlation_threshold = correlation_threshold_pkg_default)
}
\arguments{
\item{df}{The dataframe to generate the EDA report}

\item{missing_threshold}{OPTIONAL. The threshold based on which variables will be flagged as "Missing". Default value is 0.3}

\item{cardinality_threshold}{OPTIONAL. The threshold based on which variables will be flagged as "High Cardinality". Default value is 50}

\item{skewness_threshold}{OPTIONAL. The threshold based on which variables will be flagged as "Highly Skewed". Default value is 20}

\item{correlation_threshold}{OPTIONAL. The threshold based on which variables will be flagged as "Highly Correlated". Default value is 0.8}
}
\value{
df_warnings A dataframe containing all the warnings and recommendations per variable based on different threshold values
}
\description{
Returns skewness information of the dataframe
}
\examples{
get_full_warnings(df)
get_full_warnings(df, missing_threshold=0.3,cardinality_threshold=50,skewness_threshold=20,correlation_threshold=0.8)
}
