% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_info}
\alias{get_info}
\title{Returns meta information of the dataframe like Number of variables, Number of observations, Total missing percentage, Total size in memory}
\usage{
get_info(df)
}
\arguments{
\item{df}{The dataframe to generate the EDA report}
}
\value{
df_info The output dataframe containing the Property and a corresponding value
}
\description{
Returns meta information of the dataframe like Number of variables, Number of observations, Total missing percentage, Total size in memory
}
\examples{
get_info(df)
}
