% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_missing_vars}
\alias{get_missing_vars}
\title{Returns missing information of the dataframe - The distribution of variables by type (factor, integer, etc.)}
\usage{
get_missing_vars(df, missing_threshold = missing_threshold_pkg_default)
}
\arguments{
\item{df}{The dataframe to generate the EDA report}

\item{missing_threshold}{OPTIONAL. The threshold based on which variables will be flagged as "Missing". Default value is 0.3}
}
\value{
perc_missing_report The output dataframe containing the column name and the percentage of missing values
}
\description{
Returns missing information of the dataframe - The distribution of variables by type (factor, integer, etc.)
}
\examples{
get_missing_vars(df)
get_missing_vars(df, missing_threshold=0.3)
}
