% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_skewed_vars}
\alias{get_skewed_vars}
\title{Returns skewness information of the dataframe}
\usage{
get_skewed_vars(df, skewness_threshold = skewness_threshold_pkg_default)
}
\arguments{
\item{df}{The dataframe to generate the EDA report}

\item{skewness_threshold}{OPTIONAL. The threshold based on which variables will be flagged as "Highly Skewed". Default value is 20}
}
\value{
list(skewness_vars, values) A list containing two elements - a dataframe with the column name and the skewness and another named list of the variables and their skewness values.
}
\description{
Returns skewness information of the dataframe
}
\examples{
get_skewed_vars(df)
get_skewed_vars(df, skewness_threshold=20)
}
