% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_unique_vars}
\alias{get_unique_vars}
\title{Returns information of the dataframe - The variables which have a constant value}
\usage{
get_unique_vars(df)
}
\arguments{
\item{df}{The dataframe to generate the EDA report}
}
\value{
list(unique_count, values) A list containing two elements - a dataframe with the column name and the unique_count and another named list of the variables and their constant values.
}
\description{
Returns information of the dataframe - The variables which have a constant value
}
\examples{
get_unique_vars(df)
}
