% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_var_types}
\alias{get_var_types}
\title{Returns meta information of the dataframe - The distribution of variables by type (factor, integer, etc.)}
\usage{
get_var_types(df)
}
\arguments{
\item{df}{The dataframe to generate the EDA report}
}
\value{
var_summary The output dataframe containing the various data types and their corresponding frequencies (number of variables)
}
\description{
Returns meta information of the dataframe - The distribution of variables by type (factor, integer, etc.)
}
\examples{
get_var_types(df)
}
