% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_report.R
\name{profile_report}
\alias{profile_report}
\title{Generates profile reports from a DataFrame. The existing implementations are great but a little basic for serious exploratory data analysis.}
\usage{
profile_report(df, output_file_path, missing_threshold = 0.3,
  cardinality_threshold = 50, skewness_threshold = 20,
  correlation_threshold = 0.8)
}
\arguments{
\item{df}{The dataframe to generate the EDA report}

\item{output_file_path}{Provide the output file path to store the html report}

\item{missing_threshold}{Flag variables as "Missing" in case this threshold is crossed. This should be between 0 and 1}

\item{cardinality_threshold}{Flag variables as "High Cardinality" in case this threshold is crossed. This should be between 0 and 1
cardinality_threshold Flag variables as "Highly Correlated" in case this threshold is crossed. This should be between 0 and 1}
}
\value{
NULL. The output HTML file is written
}
\description{
Generates profile reports from a DataFrame. The existing implementations are great but a little basic for serious exploratory data analysis.
}
\examples{
profile_report(mtcars,"report.html")
profile_report(mtcars,"/path/to/report.html",missing_threshold=0.30,cardinality_threshold=50,skewness_threshold=20,correlation_threshold=0.8)
}
